Use Lucida Grande on all Macs for now
In 8c4552091 we started forcing Lucida Grande on non-Retina displays to
fix San Francisco looking terrible. I thought that was due to Catalina
causing San Francisco to be used and San Francisco looking terrible on
non-Retina displays, but actually it's just that San Francisco looks
terrible in Catalina, with much narrower spacing, on all displays,
likely due to a bug in Firefox 52/60. (My non-Retina system was
apparently configured to force use of Lucida Grande in some programs,
but SF is in fact used in a Mojave VM, and it looks fine.)

I haven't found a Bugzilla ticket for this yet, but there's a similar
ticket open for Chromium [1]. It's possible it's been fixed in later
versions of Firefox, so we might be able to backport a fix.

In the meantime, this isn't ideal, but it's better than
squished-together San Francisco.

[1] https://bugs.chromium.org/p/chromium/issues/detail?id=1005969